requirements.txt
psycopg2-binary==2.9.9
apache-airflow-providers-postgres>=3.0.0
